feat: add language attribute to RedisConfig (#953)
* feat: add language to RedisConfig


* test: add test for redis chinese support


* docs: add language to RedisConfig


* fix: numpy is updated to 1.24.0


* fix: increase timeout-minutes for oldproto

+++ docarray/array/storage/opensearch/find.py
@@ -166,7 +166,7 @@ class FindMixin(BaseFindMixin):
         :return: DocumentArray containing the closest documents to the query if it is a single query, otherwise a list of DocumentArrays containing
            the closest Document objects for each of the queries in `query`.
         """
-        query = np.array(query).astype(np.float)
+        query = np.array(query).astype(np.float32)
         num_rows, n_dim = ndarray.get_array_rows(query)
         if n_dim != 2:
             query = query.reshape((num_rows, -1))
